feat: scene and example usage

+++ src/context.rs
@@ -1,8 +1,8 @@
 extern "C" {
     fn js_clear_screen_to_color(red: f32, green: f32, blue: f32, alpha: f32);
     fn js_draw_rectangle(
-        x: f32,
-        y: f32,
+        x: f64,
+        y: f64,
         width: f32,
         height: f32,
         red: f32,
@@ -23,8 +23,8 @@ impl Context {
     
     pub fn draw_rectangle(
         &mut self,
-        x: f32,
-        y: f32,
+        x: f64,
+        y: f64,
         width: f32,
         height: f32,
         red: f32,
